cart implemented

+++ BusinessLogicLayer/DTOs/InstructorDTO.cs
@@ -33,7 +33,7 @@ namespace BusinessLogicLayer.DTOs
         public string InstructorImage { get; set; }
 
         [Required]
-        public int TotalIncome { get; set; }
+        public double TotalIncome { get; set; }
 
         [Required]
         public double Rating { get; set; }
